+++ src/lib/Code.svelte
@@ -9,3 +9,10 @@
 <code class="text-primary-700 dark:text-primary-700">
   {@render children()}
 </code>
+
+<!--
+@component
+[Go to docs](https://svelte-icon-webkit.codewithshin.com/)
+## Props
+@props: children: Snippet;
+-->