+++ src/lib/H2.svelte
@@ -12,3 +12,10 @@
   {@render children()}
   </Anchor
 >
+
+<!--
+@component
+[Go to docs](https://svelte-icon-webkit.codewithshin.com/)
+## Props
+@props: children: Snippet;
+-->