+++ src/lib/Nav.svelte
@@ -100,5 +100,5 @@
 @props:twitterUrl?: string;
 @props:githubUrl?: string;
 @props:headerClass?: string;
-@props:urlsToInclude?: string[];
+@props:urlsToIncludeSwitcher?: string[];
 -->