+++ src/lib/Sidebar.svelte
@@ -84,3 +84,10 @@
     
   </SidebarGroup>
 </Sidebar>
+
+<!--
+@component
+[Go to docs](https://svelte-icon-webkit.codewithshin.com/)
+## Props
+@props: 
+-->