+++ src/lib/Toc.svelte
@@ -77,5 +77,6 @@
 @component
 [Go to docs](https://svelte-icon-webkit.codewithshin.com/)
 ## Props
-@props: 
+@props: headingSelector: string;
+@props:extract: (x
 -->